requirements.txt
streamlit
streamlit-extras
python-dotenv
PyPDF2
openai
requests
typing
diagrams
graphviz
openai-agents
pandas
altair
pydantic
aiohttp
nest-asyncio
